backports: add 3.0.13_sles11_sp2 backport
This kernel differs slightly from the vanilla 3.0.x
kernel in that the netlink_dump_start() routine
has an extra arg.

To enable compilation on both kernels, this patch
forks the SLES11SP2 backport from the 3.0 backport
and adjusts the compat include via the NEED_MIN_DUMP_ALLOC_ARG
to keep the extra arg.

This patch also adjusts the RHEL6.2 backport to leave
the source as it.

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@intel.com>