ib_srp: Avoid that module removal can trigger a deadlock
Avoid that scsi_remove_host() is invoked from the context of a work
queue thread on which work has been queued that scsi_remove_host()
might be waiting for. That avoids that module removal of ib_srp
triggers a deadlock on a pre-2.6.36 kernel. This patch has been
tested on RHEL 6.1, RHEL 6.2, RHEL 6.3 and SLES 11 SP2.

Reported-by: Rupert Dance <rsdance@soft-forge.com>
Signed-off-by: Bart Van Assche <bvanassche@acm.org>