xeon-phi: ibscif bug 2575: Allow xeon-phi support for Knights Landing
New patch file for the compat-rdma tech-preview/xeon-phi directory

The Knights Landing (KNL) version of Intel mpss will use the scif driver that
was upstreamed to kernel.org.  Although the data structure layout is the same
as the Knights Corner (KNC) version, some functions have changed.  The
scif_get_nodeIDs call is renamed to scif_get_node_ids and the scif_pci_info is
removed.  These changes impact the ibscif and ib_qib drivers for KNL.  The
attached patch will allow OFED-3.18-1 --with-xeon-phi to work with both
versions of mpss.  This new patch file should be added to the compat-rdma
source rpm in the tech-preview/xeon-phi directory.

Signed-off-by: Jay Sternberg <jay.e.sternberg@intel.com>