bnxt_re: BZ 2654 Fix traffic failure after changing ip address
While adding a new gid, the driver currently does not return the context
back to the stack. A subsequent del_gid() (e.g, when ip address is changed)
doesn't find the right context in the driver and it ends up dropping that
request. This results in the HW caching a stale gid entry and traffic fails
because of that. Fix by returning the proper context in bnxt_re_add_gid()

Signed-off-by: Selvin Xavier <selvin.xavier@broadcom.com>